# docs/openapi.yaml
openapi: "3.1.0"
...
info:
  title: Chronos Ledger API
  version: "0.10.0" # x-release-please-version
  description: |
    Organization Schedule & Attendance Management REST API.

    **Authentication**: All endpoints except `/config` and `/auth/login`
    require a Bearer JWT in the `Authorization` header. An API key in
    `X-API-Key` is accepted in place of one anywhere, which is how a lobby
    terminal with no one to sign in reaches `/guest/register-checkin` and
    `/guest/directory`. A key can be narrowed to part of the API and given
    an end date: see the `ApiKeyAuth` security scheme.

    **Roles**: `SUPER_ADMIN` > `UNIT_ADMIN` > `STAFF` > `MEMBER`.
    Role requirements are noted on each endpoint.
  contact:
    name: Chronos Ledger Contributors
    email: admin@org.internal
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost
    description: Local development (Docker Compose)
  - url: http://{server_ip}
    description: Organization intranet
    variables:
      server_ip:
        default: "192.168.1.100"

# ─────────────────────────────────────────────
# Reusable components
# ─────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A machine credential bound to a service-account user. Accepted anywhere
        a Bearer token is, and acts as the bound user, so roles and unit scoping
        apply unchanged. This is how a lobby kiosk authenticates.

        A key can also be narrowed and given an end date, which a Bearer token
        cannot be. `scopes` is a comma separated list of `<area>:read` and
        `<area>:write`, where the area is the path segment after `/api/v1`:
        `schedule`, `resources`, `attendance`, `users`, `ingestion`, `guest`,
        `sync`, `config`, `auth`, `api-keys`. `GET`, `HEAD` and `OPTIONS` are
        reads and everything else is a write, so `schedule:read` passes
        `GET /schedule/slots` and is refused `POST /schedule/slots` with a 403
        naming `schedule:write`. The single scope `*` is every area, which is
        what a key issued before scopes existed holds and what a key created
        without naming any gets.

        `expires_at` is checked on the request, not by a sweep, so a key stops
        working the moment it runs out and answers 401 `API key expired`. A
        null expiry never runs out.

        The first-login gate does not apply to a key. It holds an admin
        account until the password it was handed is changed, and a key never
        uses that password, so a service account at an admin role works
        through its key without anyone signing in as it.
  schemas:
...
    LogVerificationState:
      type: string
      enum: [PENDING_VERIFICATION, VERIFIED_APPROVED, VERIFIED_DENIED]
      description: Approval workflow state for absence requests and guest handshakes.

    # ── Auth ──────────────────────────────────
...
    FeedToken:
      type: object
      properties:
        feed_token:
          type: string
          description: 256 bits out of the system random source, hex encoded.
        feed_path:
          type: string
          description: >
            The feed path this token opens, relative to the server root.
          example: /api/v1/sync/user-feed/1f4a9c2e.ics

    # ── Users ─────────────────────────────────
...
    GuestCheckInRequest:
      type: object
      required: [guest_name, contact_phone, originating_body, target_staff_id, visitation_intent]
      properties:
        guest_name:
          type: string
          minLength: 2
          maxLength: 100
          example: John Smith
        contact_phone:
          type: string
          minLength: 8
          maxLength: 20
          pattern: "^[-0-9+() ]+$"
          example: "+91 98765 43210"
        originating_body:
          type: string
          minLength: 2
          maxLength: 100
          example: TechCorp Ltd
        target_staff_id:
          type: string
          minLength: 1
          maxLength: 50
          example: FAC001
        visitation_intent:
          type: string
          minLength: 10
          maxLength: 500
          example: Research collaboration discussion
    GuestDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          $ref: "#/components/schemas/LogVerificationState"
    GuestResponse:
      type: object
      properties:
        id:
          type: integer
        guest_name:
          type: string
        contact_phone:
          type: string
        originating_body:
          type: string
        target_staff_id:
          type: string
        visitation_intent:
          type: string
        handshake_status:
          $ref: "#/components/schemas/LogVerificationState"
        timestamp_marked:
          type: string
          format: date-time
    StaffAvailabilityResponse:
      type: object
      properties:
        staff_id:
          type: string
        full_name:
          type: string
        unit_code:
          type: string
          nullable: true
        availability_label:
          type: string
          description: Human-readable availability derived from current_occupancy_index.

    # ── Common ────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Not authorized.
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but insufficient role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: >
        The caller is past its budget for this route. Sized by the
        RATE_LIMIT_* settings; the counters live in Redis, and an instance
        that cannot reach Redis stops applying the limits rather than
        refusing the requests.
      headers:
        Retry-After:
          description: Seconds until the budget refills.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

# Either credential is accepted on every endpoint below.
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ─────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────
paths:
...
  /api/v1/guest/register-checkin:
    post:
      tags: [Guest Gate]
      summary: Register an organization visitor (kiosk)
      description: >
        The visitor does not authenticate; the kiosk device does, with an
        admin-issued API key in `X-API-Key`.
        Fires a `GUEST_HANDSHAKE_REQ` WebSocket event to the target staff member.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestCheckInRequest"
      responses:
        "201":
          description: Guest registered. Staff notified via WebSocket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestResponse"
        "404":
          description: Target staff not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/guest/directory:
    get:
      tags: [Guest Gate]
      summary: Staff availability directory (kiosk)
      description: >
        Returns staff with OPEN_AD_HOC or VERY_FREE status for kiosk display.
        Requires a credential: this response carries the staff roster and each
        person's live presence.
      parameters:
        - name: name
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 100
          description: >
            Case-insensitive search by staff name. Two characters minimum, so
            the roster cannot be walked one letter at a time.
      responses:
        "200":
          description: Available staff.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffAvailabilityResponse"
...
  /api/v1/guest/{entry_id}/decide:
    patch:
      tags: [Guest Gate]
      summary: Accept or decline a guest visit
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestDecisionRequest"
      responses:
        "200":
          description: Decision recorded. If approved, guest receives notification.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Ingestion ─────────────────────────────
...
  /api/v1/sync/feed-token:
    get:
      tags: [Calendar Sync]
      summary: Get the caller's calendar feed token
      description: >
        Returns the token that keys the caller's iCalendar feed, minting one on
        the first call. `feed_path` is the path to subscribe, already built from
        the token. The token is the only thing guarding the feed, so hand it out
        the way a password is handed out.
      responses:
        "200":
          description: The feed token and the path built from it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/sync/feed-token/rotate:
    post:
      tags: [Calendar Sync]
      summary: Replace the caller's calendar feed token
      description: >
        Mints a new token and invalidates the old one. Every calendar app
        subscribed to the old path stops updating and has to be resubscribed,
        which is the point: this is what to call when a feed URL has leaked.
      responses:
        "200":
          description: The new feed token and the path built from it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/sync/user-feed/{feed_token}.ics:
    get:
      tags: [Calendar Sync]
      summary: iCalendar feed for a user (rolling 37-day window)
      description: >
        Subscribe this URL directly in Google Calendar, Apple Calendar, or Outlook.
        Returns standard iCalendar (text/calendar) content.
        Staff feed includes all sessions where they are active or substitute lead.
        Member feed includes all registered activities.
        Timed sessions are written as UTC instants (DTSTART and DTEND end in Z),
        converted from the organization's wall clock through ORG_TIMEZONE, so a
        subscriber reading the feed from anywhere sees the session at the moment
        it happens. An entry with no window of its own is an all-day date and
        carries no zone, because a date reads the same everywhere. That is an
        ad-hoc day nobody gave a time to, or a day generated before the ledger
        carried its own window.
      parameters:
        - name: feed_token
          in: path
          required: true
          schema:
            type: string
          description: >
            The token from `GET /api/v1/sync/feed-token`, not a user id. The
            feed itself is unauthenticated so a calendar app can poll it, which
            leaves the token as the only thing guarding it.
      responses:
        "200":
          description: iCalendar feed.
          content:
            text/calendar:
              schema:
                type: string
                example: "BEGIN:VCALENDAR\nVERSION:2.0\n..."
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ── WebSocket ─────────────────────────
...